% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersection}
\alias{intersection}
\title{A intersection function}
\usage{
intersection(...)
}
\arguments{
\item{...}{Vector with master values to check and vector to compare values against}
}
\value{
Character vector of all common attributes
}
\description{
\code{intersection()} returns all intersections it found.
}
\examples{
input1 <- seq(50, 100, by=10)
input2 <- seq(70, 130, by=10)
intersection(input1, input2)
}
\keyword{intersect,}
\keyword{intersection}
