% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_intersect}
\alias{list_intersect}
\title{A list_intersect function}
\usage{
list_intersect(...)
}
\arguments{
\item{vect1}{Vector of string to make comparison by.}

\item{vect2}{Vector of string to make comparison against.}
}
\value{
Character vector of all common attributes
}
\description{
\code{list_intersect()} returns all intersections it found.
}
\examples{
numb1 <- seq(50, 100, by=10)
numb2 <- seq(70, 130, by=10)
list_intersect(numb1, numb2)
}
\keyword{intersect}
