% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisetDemo-data.R
\docType{data}
\name{mulsetDemo}
\alias{mulsetDemo}
\title{Demo data from mulset package.
Used in examples, regarding missing value columns.}
\format{An object of class \code{data.frame} with 4 rows and 7 columns.}
\usage{
data(mulsetDemo)
}
\description{
Demo data from mulset package.
Used in examples, regarding missing value columns.
}
\examples{
data(mulsetDemo)
resamples <- mulset(mulsetDemo, exclude = c("outcome", "age", "gender"), 250)

}
\keyword{datasets}
