% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mulset}
\alias{mulset}
\title{A mulset function}
\usage{
mulset(data, exclude = c("outcome", "age", "gender"),
  maxIntersections = NULL)
}
\arguments{
\item{data}{Data-frame containing your data}

\item{exclude}{List of columns to exclude from analysis}

\item{maxIntersections}{Maximum number of unique multi-set intersections to generate}
}
\value{
If any intersections around found it returns a list that contains all available multi-sets
}
\description{
\code{mulset()} returns all multi-set intersections it found.
}
\details{
This function allows you to generate specific type of multi-set intersections.
}
\examples{

data <- data.table::fread("../data/Example_100.csv", header = T, sep = ',', stringsAsFactors = FALSE, data.table = FALSE)
resamples <- mulset(data, exclude = c("outcome", "age", "gender"), 250)

}
\keyword{mulset}
