% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mulset}
\alias{mulset}
\title{A mulset function}
\usage{
mulset(data, exclude = NULL, include = c("samples", "samples_count",
  "datapoints"), maxIntersections = NULL, hashMethod = "md5",
  names = FALSE)
}
\arguments{
\item{data}{Data frame containing your incomplete data.}

\item{exclude}{Vector  of  strings  containing  one  or  more  variable  names  from names(data)}

\item{include}{List of attributes to return in results. Possible values are: c("samples", "samples_count", "datapoints"). If parameter is set to NULL only features will be returned.}

\item{maxIntersections}{Maximum number of unique multi-set intersections to generate}

\item{hashMethod}{Hashing method to use for unique sets identification. Available choices: md5 (default),sha1,crc32,sha256,sha512,xxhash32,xxhash64,murmur32}

\item{names}{Should we reset return list keys or keep original features hash as a key value}
}
\value{
If any intersections are found it returns a list that contains all available multi-set intersections
You can convert this to data-frame following example provided or use it as it is.
}
\description{
\code{mulset()} returns all multi-set intersections it found.
}
\details{
This function allows you to generate specific type of multi-set intersections.
It searches for multi set intersections between rows and column identifiers.
}
\examples{
data(mulsetDemo)
resamples <- mulset(mulsetDemo, exclude = c("outcome", "age", "gender"), 250)

resamplesFrame <- as.data.frame(t(sapply(resamples,c)))
}
\keyword{data}
\keyword{intersection,}
\keyword{missing}
\keyword{mulset,}
\keyword{multi-set}
\keyword{table}
