% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisetDemo-data.R
\docType{data}
\name{mulsetDemo}
\alias{mulsetDemo}
\title{Demo data set from mulset package.
This data is used in this package examples. It consist of 4x4 feature matrix + additional dummy columns that can be used for testing.}
\format{An object of class \code{data.frame} with 4 rows and 7 columns.}
\usage{
data(mulsetDemo)
}
\description{
Demo data set from mulset package.
This data is used in this package examples. It consist of 4x4 feature matrix + additional dummy columns that can be used for testing.
}
\examples{
data(mulsetDemo)
print(head(mulsetDemo))
resamples <- mulset(mulsetDemo, exclude = c("outcome", "age", "gender"))

}
\keyword{datasets}
